fix warning in test, add/modify implementations
- modify `c89atomic_is_lock_free` and bug fix Linux usage was giving `undefined reference to `__atomic_is_lock_free'`
- using `MemoryBarrier()`, `_ReadWriteBarrier()` for Windows for `atomic_thread_fence` and  `atomic_signal_fence`
- Additional usage reference
- change atomic_compare_exchange_* macros to `atomic_cas` funcs

+++ tests/tests-core-atomic.c
@@ -54,11 +54,11 @@ static void test_atomic_inc_rollover(void) {
         CHK_EXPECTED(INT_MAX - 30 + i + 1, atomic_load(&res));
     }
     CHK_EXPECTED(INT_MAX, atomic_fetch_add(&res, 1));
-    CHK_EXPECTED(INT_MIN, atomic_load(&res));
-    CHK_EXPECTED(INT_MIN, atomic_fetch_add(&res, 1));
-    CHK_EXPECTED(INT_MIN + 1, atomic_load(&res));
-    CHK_EXPECTED(INT_MIN + 1, atomic_fetch_add(&res, 1));
-    CHK_EXPECTED(INT_MIN + 2, atomic_load(&res));
+    CHK_EXPECTED(INT_MIN, (int)atomic_load(&res));
+    CHK_EXPECTED(INT_MIN, (int)atomic_fetch_add(&res, 1));
+    CHK_EXPECTED(INT_MIN + 1, (int)atomic_load(&res));
+    CHK_EXPECTED(INT_MIN + 1, (int)atomic_fetch_add(&res, 1));
+    CHK_EXPECTED(INT_MIN + 2, (int)atomic_load(&res));
 }
 
 /* Test atomic_fetch_sub */
@@ -99,11 +99,11 @@ static void test_atomic_dec_rollover(void) {
         CHK_EXPECTED(INT_MIN + 30 - i - 1, atomic_load(&res));
     }
     CHK_EXPECTED(INT_MIN, atomic_fetch_sub(&res, 1));
-    CHK_EXPECTED(INT_MAX, atomic_load(&res));
-    CHK_EXPECTED(INT_MAX, atomic_fetch_sub(&res, 1));
-    CHK_EXPECTED(INT_MAX - 1, atomic_load(&res));
-    CHK_EXPECTED(INT_MAX - 1, atomic_fetch_sub(&res, 1));
-    CHK_EXPECTED(INT_MAX - 2, atomic_load(&res));
+    CHK_EXPECTED(INT_MAX, (int)atomic_load(&res));
+    CHK_EXPECTED(INT_MAX, (int)atomic_fetch_sub(&res, 1));
+    CHK_EXPECTED(INT_MAX - 1, (int)atomic_load(&res));
+    CHK_EXPECTED(INT_MAX - 1, (int)atomic_fetch_sub(&res, 1));
+    CHK_EXPECTED(INT_MAX - 2, (int)atomic_load(&res));
 }
 
 /* Test atomic_cas with a correct old value */
